email-open-discord-notifier/requirements.txt
# Email Open Discord Notifier Dependencies
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
